+++ src/components/library/Library.jsx
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { Outlet } from 'react-router'
 
 import { loadWorkTypes } from 'actions/library'
-import { Tab, Tabs } from 'components/generics/Tab'
+import { Tab, TabBar } from 'components/generics/TabBar'
 import { Status } from 'reducers/alterationsResponse'
 import { workTypeStatePropType } from 'reducers/library'
 
@@ -41,11 +41,11 @@ class Library extends Component {
 
     return (
       <div id="library" className="box">
-        <Tabs>
+        <TabBar>
           <Tab to="/library/song" iconName="music" extraClassName="home" />
           <Tab to="/library/artist" iconName="microphone-alt" name="Artists" />
           {workTypesTabs}
-        </Tabs>
+        </TabBar>
         <Outlet />
       </div>
     )
