+++ src/components/playlist/Playlist.jsx
@@ -1,13 +1,13 @@
 import { Component } from 'react'
 import { Outlet } from 'react-router'
 
-import { Tab, Tabs } from 'components/generics/Tab'
+import { Tab, TabBar } from 'components/generics/TabBar'
 
 export default class Playlist extends Component {
   render() {
     return (
       <div id="playlist" className="box">
-        <Tabs>
+        <TabBar>
           <Tab
             to="/playlist/queueing"
             iconName="chevron-right"
@@ -19,7 +19,7 @@ export default class Playlist extends Component {
             iconName="exclamation-triangle"
             name="Errors"
           />
-        </Tabs>
+        </TabBar>
         <Outlet />
       </div>
     )
