+++ src/components/settings/Settings.jsx
@@ -1,13 +1,13 @@
 import { Component } from 'react'
 import { Outlet } from 'react-router'
 
-import { Tab, Tabs } from 'components/generics/Tab'
+import { Tab, TabBar } from 'components/generics/TabBar'
 
 export default class Settings extends Component {
   render() {
     return (
       <div id="settings" className="box">
-        <Tabs>
+        <TabBar>
           <Tab to="/settings/users" iconName="users" name="Users" />
           <Tab to="/settings/song-tags" iconName="tags" name="Tags" />
           <Tab to="/settings/kara-status" iconName="play" name="Kara status" />
@@ -17,7 +17,7 @@ export default class Settings extends Component {
             name="Kara stop time"
           />
           <Tab to="/settings/tokens" iconName="user-circle" name="Tokens" />
-        </Tabs>
+        </TabBar>
         <Outlet />
       </div>
     )
