+++ src/style/components/generics/_tab_bar.scss
@@ -8,26 +8,23 @@
 @use '~/abstracts/sizes';
 @use '~/abstracts/support';
 
-// tab bar sizes
-$tab-bar-height: sizes.$row-height;
-$tab-bar-icon-font-size: 1.5em;
-
-// `tabs` class:
+// `tab-bar` class:
 //
-// The `tabs` class is aimed to stylize a tab bar, contiaining several tabs of
-// class `tab`. Tabs are actually controls with extra styling.  The active tab
-// (with the `active` class) has a dark background.
+// The `tab-bar` class is aimed to stylize a tab bar, contiaining several tabs
+// of class `tab`. Tabs are actually controls with extra styling. The active
+// tab (with the `active` class) has a dark background.
 //
 // Each tab can have a `icon` span and/or a `name` span, for its icon and text.
 //
-// By default, each tab has the size of its content and the tab bar has the size
-// of all the tabs. To obtain a tab bar where the tabs are stretched on all the
-// space available, add the `flex: 1;` property to the element with the `tab`
-// class.
+// By default, each tab has the size of its content and the tab bar has the
+// size of all the tabs. To obtain a tab bar where the tabs are stretched on
+// all the space available, add the `flex: 1;` property to the element with the
+// `tab` class.
 //
-// Usually, the `tab-bar` element is a `nav` and the `tab` a `a` (`Link` for
-// React).
-.tabs {
+// Usually, the `tab-bar` element is a `nav` and the `tab` an `a` (`NavLink`
+// for React). `NavLin` is always rendered as an `a`, even when active. This is
+// why there are some styling for when hovered or focused.
+.tab-bar {
   justify-content: flex-start;
   overflow-x: scroll;
 
@@ -51,14 +48,18 @@ $tab-bar-icon-font-size: 1.5em;
     }
 
     &.active {
+      // override color when active
+      // must use !importatn as even controls have a higher priority
       background: colors.$neutral-110 !important;
       color: colors.$neutral-410 !important;
       cursor: default;
 
+      // make it hoverizable, as it is always rendered as an `a`
       &:hover:not(:focus) {
         background: colors.$neutral-112 !important;
       }
 
+      // make it focusable, as it is always rendered as an `a`
       &:focus {
         // XXX wrong color
         background: colors.$neutral-112 !important;
